Adicionado classes model para Campanha,Produto e Usuario
