+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".MainActivity"
+            android:name=".telas.MainActivity"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar">
             <intent-filter>
@@ -22,8 +22,8 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".ActCadastro" />
-        <activity android:name=".ActQrCode"
+        <activity android:name=".telas.ActCadastro" />
+        <activity android:name=".telas.ActQrCode"
             android:screenOrientation="portrait"
             android:configChanges="orientation"></activity>
