+++ app/src/main/java/br/ufg/inf/dsdm/kleudson/wififacil/model/Produto.java
@@ -1,0 +1,31 @@
+package br.ufg.inf.dsdm.kleudson.wififacil.model;
+
+/**
+ * Created by Jean on 11/20/16.
+ */
+
+public class Produto {
+    private String nome;
+    private double preco;
+
+    public Produto(String nome, double preco) {
+        this.nome = nome;
+        this.preco = preco;
+    }
+
+    public String getNome() {
+        return nome;
+    }
+
+    public void setNome(String nome) {
+        this.nome = nome;
+    }
+
+    public double getPreco() {
+        return preco;
+    }
+
+    public void setPreco(double preco) {
+        this.preco = preco;
+    }
+}
