+++ app/src/main/java/br/ufg/inf/dsdm/kleudson/wififacil/model/Usuario.java
@@ -1,0 +1,31 @@
+package br.ufg.inf.dsdm.kleudson.wififacil.model;
+
+/**
+ * Created by Jean on 11/20/16.
+ */
+
+public class Usuario {
+    private String nome;
+    private String senha;
+
+    public Usuario(String nome, String senha) {
+        this.nome = nome;
+        this.senha = senha;
+    }
+
+    public String getNome() {
+        return nome;
+    }
+
+    public void setNome(String nome) {
+        this.nome = nome;
+    }
+
+    public String getSenha() {
+        return senha;
+    }
+
+    public void setSenha(String senha) {
+        this.senha = senha;
+    }
+}
