+++ app/src/main/res/layout/act_cadastro.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context="br.ufg.inf.dsdm.kleudson.wififacil.ActCadastro"
+    tools:context="br.ufg.inf.dsdm.kleudson.wififacil.telas.ActCadastro"
     android:weightSum="1"
     >
