+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="br.ufg.inf.dsdm.kleudson.wififacil.MainActivity">
+    tools:context="br.ufg.inf.dsdm.kleudson.wififacil.telas.MainActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
