+++ app/src/main/res/layout/content_main.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="br.ufg.inf.dsdm.kleudson.wififacil.MainActivity"
+    tools:context="br.ufg.inf.dsdm.kleudson.wififacil.telas.MainActivity"
     tools:showIn="@layout/activity_main">
 
     <com.google.android.gms.ads.AdView
