+++ app/src/main/res/menu/menu_bottom.xml
@@ -2,7 +2,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".MainActivity">
+    tools:context=".telas.MainActivity">
     <item
         android:id="@+id/action_salvar"
         android:orderInCategory="100"
